% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import an R File or Module}
\usage{
import(file, modulize = F)
}
\arguments{
\item{file}{A character string specifying the path to the R script file to be imported.}

\item{modulize}{A logical value indicating whether to convert the file into a module structure.
If `TRUE`, the function uses `buildModule` to process the imported file.
Defaults to `FALSE`.}
}
\value{
A list representation of the imported file's environment. If `modulize = TRUE`, the list
represents the module structure created by `buildModule`.
}
\description{
This function imports an R script file into the current session, optionally converting it into a module structure.
}
\examples{
# Import an R script without modulizing
env = import(fileWhenTest('import.R'))

# Import an R script and convert it into a module
module = import(fileWhenTest('import.R'), modulize = TRUE)

}
