% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withAssume}
\alias{withAssume}
\title{withAssume}
\usage{
withAssume(expr, onPass = "Success", onError = "Error", ...)
}
\arguments{
\item{expr}{raw expression}

\item{...}{Temporary variable}
}
\description{
Print message after judging the expression exit status.
}
