% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withClusters}
\alias{withClusters}
\title{withClusters}
\usage{
withClusters(expr, nWorker = NULL, ...)
}
\arguments{
\item{expr}{raw expression}

\item{nWorker}{n of sessions}

\item{...}{Temporary variable}
}
\description{
First open multiple threads, then close.
}
