% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withOption}
\alias{withOption}
\title{withOption}
\usage{
withOption(expr, options, ...)
}
\arguments{
\item{expr}{raw expression}

\item{options}{a list. options.}

\item{...}{Temporary variable}
}
\description{
First change the options when excuting expression, then change it back.
}
