% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withSessions}
\alias{withSessions}
\title{withSessions}
\usage{
withSessions(expr, nWorker = 4, header = "", ...)
}
\arguments{
\item{expr}{raw expression}

\item{nWorker}{n of sessions}

\item{header}{to source}

\item{...}{Temporary variable}
}
\description{
First open multiple threads, then close.
}
