% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withUnlocking}
\alias{withUnlocking}
\title{withUnlocking}
\usage{
withUnlocking(expr, old, new, ...)
}
\arguments{
\item{expr}{raw expression}

\item{old}{old function}

\item{new}{new function}

\item{...}{Temporary variable}
}
\description{
First replace function in specific namespace, then recover
}
